Created new pyro submodule and moved PyroLikelihood classes
